#f/148254-adaptacion-crud-medicos

See merge request desarrollo/1187-001-hope/hopes-back!14

+++ src/main/java/es/plexus/hopes/hopesback/controller/DoctorController.java
@@ -5,6 +5,9 @@ import es.plexus.hopes.hopesback.service.DoctorService;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.Pageable;
+import org.springframework.data.web.PageableDefault;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -18,13 +21,14 @@ import org.springframework.web.servlet.support.ServletUriComponentsBuilder;
 
 import javax.validation.Valid;
 import java.net.URI;
-import java.util.List;
+import java.util.Objects;
 
 @RestController
 @RequestMapping("/doctor")
 public class DoctorController {
 
 	private static final Logger LOGGER = LogManager.getLogger(DoctorController.class);
+	private static final String CALLING_SERVICE = "Calling service...";
 
 	private final DoctorService doctorService;
 
@@ -34,9 +38,9 @@ public class DoctorController {
 	}
 
 	@GetMapping
-	public ResponseEntity<List<DoctorDTO>> getAllDoctors() {
-		LOGGER.debug("Calling service...");
-		return ResponseEntity.ok(doctorService.getAllDoctors());
+	public ResponseEntity<Page<DoctorDTO>> getAllDoctors(@PageableDefault(size = 5) Pageable pageable) {
+		LOGGER.debug(CALLING_SERVICE);
+		return ResponseEntity.ok(doctorService.getAllDoctors(pageable));
 	}
 
 	@GetMapping("/{id}")
@@ -46,7 +50,7 @@ public class DoctorController {
 
 	@PostMapping
 	public ResponseEntity<DoctorDTO> addDoctor(@RequestBody @Valid final DoctorDTO doctorDTO) {
-		LOGGER.debug("Calling service...");
+		LOGGER.debug(CALLING_SERVICE);
 		final DoctorDTO doctor = doctorService.addDoctor(doctorDTO);
 		final URI location = ServletUriComponentsBuilder
 				.fromCurrentRequest()
@@ -57,17 +61,29 @@ public class DoctorController {
 		return ResponseEntity.created(location).body(doctor);
 	}
 
-	@PutMapping("/{id}")
-	public ResponseEntity<DoctorDTO> updateDoctor(@PathVariable Long id, @RequestBody @Valid final DoctorDTO doctorDTO) {
-		doctorDTO.setId(id);
-		LOGGER.debug("Calling service...");
+	@PutMapping
+	public ResponseEntity<DoctorDTO> updateDoctor(@RequestBody @Valid final DoctorDTO doctorDTO) {
+		final Long id = doctorDTO.getId();
+		if (checkDoctorExistence(id)) return ResponseEntity.badRequest().build();
+
+		LOGGER.debug(CALLING_SERVICE);
 		return ResponseEntity.ok(doctorService.updateDoctor(doctorDTO));
 	}
 
-	@DeleteMapping("/{id}")
-	public void deleteDoctor(@PathVariable Long id) {
-		LOGGER.debug("Calling service...");
+	@DeleteMapping
+	public ResponseEntity<Object> deleteDoctor(@RequestBody Long id) {
+		if (checkDoctorExistence(id)) return ResponseEntity.badRequest().build();
+		LOGGER.debug(CALLING_SERVICE);
 		doctorService.deleteDoctor(id);
+
+		return ResponseEntity.ok().build();
 	}
 
+	private boolean checkDoctorExistence(Long id) {
+		if (Objects.isNull(doctorService.getOneDoctor(id))) {
+			LOGGER.error("Not found doctor with id=" + id);
+			return true;
+		}
+		return false;
+	}
 }
